fix(derive): refactor tx enveloped
