+++ crates/derive/src/types/batch/span_batch/tx_data/eip1559.rs
@@ -1,7 +1,10 @@
 //! This module contains the eip1559 transaction data type for a span batch.
 
 use crate::types::eip2930::AccessList;
-use alloy_primitives::U256;
+use crate::types::{
+    network::Signed, SpanBatchError, SpanDecodingError, Transaction, TxEip1559, TxEnvelope, TxKind,
+};
+use alloy_primitives::{Address, Signature, U256};
 use alloy_rlp::{Bytes, Decodable, Encodable, Header};
 
 /// The transaction data for an EIP-1559 transaction within a span batch.
@@ -17,6 +20,49 @@ pub struct SpanBatchEip1559TransactionData {
     pub data: Bytes,
     /// Access list, used to pre-warm storage slots through static declaration.
     pub access_list: AccessList,
+}
+
+impl SpanBatchEip1559TransactionData {
+    /// Converts [SpanBatchEip1559TransactionData] into a [TxEnvelope].
+    pub fn to_enveloped_tx(
+        &self,
+        nonce: u64,
+        gas: u64,
+        to: Option<Address>,
+        chain_id: u64,
+        signature: Signature,
+    ) -> Result<TxEnvelope, SpanBatchError> {
+        let eip1559_tx = TxEip1559 {
+            chain_id,
+            nonce,
+            max_fee_per_gas: u128::from_be_bytes(
+                self.max_fee_per_gas.to_be_bytes::<32>()[16..]
+                    .try_into()
+                    .map_err(|_| {
+                        SpanBatchError::Decoding(SpanDecodingError::InvalidTransactionData)
+                    })?,
+            ),
+            max_priority_fee_per_gas: u128::from_be_bytes(
+                self.max_priority_fee_per_gas.to_be_bytes::<32>()[16..]
+                    .try_into()
+                    .map_err(|_| {
+                        SpanBatchError::Decoding(SpanDecodingError::InvalidTransactionData)
+                    })?,
+            ),
+            gas_limit: gas,
+            to: if let Some(to) = to {
+                TxKind::Call(to)
+            } else {
+                TxKind::Create
+            },
+            value: self.value,
+            input: self.data.clone().into(),
+            access_list: self.access_list.clone(),
+        };
+        let signature_hash = eip1559_tx.signature_hash();
+        let signed_eip1559_tx = Signed::new_unchecked(eip1559_tx, signature, signature_hash);
+        Ok(TxEnvelope::Eip1559(signed_eip1559_tx))
+    }
 }
 
 impl Encodable for SpanBatchEip1559TransactionData {
