+++ crates/derive/src/types/batch/span_batch/tx_data/wrapper.rs
@@ -4,10 +4,7 @@ use super::{
     SpanBatchEip1559TransactionData, SpanBatchEip2930TransactionData,
     SpanBatchLegacyTransactionData,
 };
-use crate::types::{
-    network::Signed, SpanBatchError, SpanDecodingError, Transaction, TxEip1559, TxEip2930,
-    TxEnvelope, TxKind, TxLegacy,
-};
+use crate::types::{SpanBatchError, SpanDecodingError, Transaction, TxEnvelope};
 use alloy_primitives::{Address, Signature, U256};
 use alloy_rlp::{Bytes, Decodable, Encodable};
 
@@ -126,89 +123,9 @@ impl SpanBatchTransactionData {
         signature: Signature,
     ) -> Result<TxEnvelope, SpanBatchError> {
         match self {
-            Self::Legacy(data) => {
-                let legacy_tx = TxLegacy {
-                    chain_id: Some(chain_id),
-                    nonce,
-                    gas_price: u128::from_be_bytes(
-                        data.gas_price.to_be_bytes::<32>()[16..]
-                            .try_into()
-                            .map_err(|_| {
-                                SpanBatchError::Decoding(SpanDecodingError::InvalidTransactionData)
-                            })?,
-                    ),
-                    gas_limit: gas,
-                    to: if let Some(to) = to {
-                        TxKind::Call(to)
-                    } else {
-                        TxKind::Create
-                    },
-                    value: data.value,
-                    input: data.data.clone().into(),
-                };
-                let signature_hash = legacy_tx.signature_hash();
-                let signed_legacy_tx = Signed::new_unchecked(legacy_tx, signature, signature_hash);
-                Ok(TxEnvelope::Legacy(signed_legacy_tx))
-            }
-            Self::Eip2930(data) => {
-                let access_list_tx = TxEip2930 {
-                    chain_id,
-                    nonce,
-                    gas_price: u128::from_be_bytes(
-                        data.gas_price.to_be_bytes::<32>()[16..]
-                            .try_into()
-                            .map_err(|_| {
-                                SpanBatchError::Decoding(SpanDecodingError::InvalidTransactionData)
-                            })?,
-                    ),
-                    gas_limit: gas,
-                    to: if let Some(to) = to {
-                        TxKind::Call(to)
-                    } else {
-                        TxKind::Create
-                    },
-                    value: data.value,
-                    input: data.data.clone().into(),
-                    access_list: data.access_list.clone(),
-                };
-                let signature_hash = access_list_tx.signature_hash();
-                let signed_access_list_tx =
-                    Signed::new_unchecked(access_list_tx, signature, signature_hash);
-                Ok(TxEnvelope::Eip2930(signed_access_list_tx))
-            }
-            Self::Eip1559(data) => {
-                let eip1559_tx = TxEip1559 {
-                    chain_id,
-                    nonce,
-                    max_fee_per_gas: u128::from_be_bytes(
-                        data.max_fee_per_gas.to_be_bytes::<32>()[16..]
-                            .try_into()
-                            .map_err(|_| {
-                                SpanBatchError::Decoding(SpanDecodingError::InvalidTransactionData)
-                            })?,
-                    ),
-                    max_priority_fee_per_gas: u128::from_be_bytes(
-                        data.max_priority_fee_per_gas.to_be_bytes::<32>()[16..]
-                            .try_into()
-                            .map_err(|_| {
-                                SpanBatchError::Decoding(SpanDecodingError::InvalidTransactionData)
-                            })?,
-                    ),
-                    gas_limit: gas,
-                    to: if let Some(to) = to {
-                        TxKind::Call(to)
-                    } else {
-                        TxKind::Create
-                    },
-                    value: data.value,
-                    input: data.data.clone().into(),
-                    access_list: data.access_list.clone(),
-                };
-                let signature_hash = eip1559_tx.signature_hash();
-                let signed_eip1559_tx =
-                    Signed::new_unchecked(eip1559_tx, signature, signature_hash);
-                Ok(TxEnvelope::Eip1559(signed_eip1559_tx))
-            }
+            Self::Legacy(data) => data.to_enveloped_tx(nonce, gas, to, chain_id, signature),
+            Self::Eip2930(data) => data.to_enveloped_tx(nonce, gas, to, chain_id, signature),
+            Self::Eip1559(data) => data.to_enveloped_tx(nonce, gas, to, chain_id, signature),
         }
     }
 }
